\name{VdgRsm-package}
\alias{VdgRsm-package}
\docType{package}
\title{
Variance Dispersion and Fraction of Design Space Plots
}
\description{
This package creates variance dispersion graphs (VDGs), fraction of design space (FDS) plots for second-order model designs in spherical and cuboidal regions, and also generates central composite and factorial designs.
}
\details{\tabular{ll}{
Package: \tab VdgRsm\cr
Type: \tab Package\cr
Version: \tab 1.02\cr
Date: \tab 2014-10-11\cr
License: \tab GPL-2
}}
\author{
Patchanok Srisuradetchai, John J Borkowski \cr
Maintainer: Patchanok Srisuradetchai <s.patchanok@gmail.com>
}
\references{
1. Borkowski J.J. (2005), Chapter 14: "Graphical Methods for Assessing the Prediction Capability of Response Surface Designs" In Khuri, A.I., "Response Surface Methodology and Related Topics", p.349-375, World Scientific Publishing.\cr
2. Giovannitti-Jensen, A. and Myers, R.H. (1989), "Graphical Assessment of the Prediction Capability of Response Surface Designs", Technometrics, 31, p.159-171.\cr
3. Khuri, A.I., Kim, H.J., and Um Y. (1996), "Quantile plots of the prediction variance for response surface designs", Computational Statistics and Data Analysis, 22, p.395-407. \cr
4. Nguyen, N.K. and Borkowski, J.J. (2008), "New 3-Level Response Surface Designs Constructed from Incomplete Block Designs", Journal of Statistical Planning and Inference, 138, p.294-305.\cr 
5. Rozum, M.A. and Myers, R.H. (1991), "Adaptation of Variance Dispersion Graphs to Cuboidal Regions of Interest", Presented at Joint Statistical Meetings, American Statistical Association, Atlanta, GA. \cr
6. Zahran, A., Anderson-Cook, C.M., and Myers, R.H. (2003), "Fraction of Design Space to Assess the Prediction Capability of Response Surface Designs", Journal of Quality Technology, 35, p.377-386.
}
\keyword{design}
\seealso{
The CRAN task view on Design of Experiments
}
