\name{Borkowski2003}
\alias{Borkowski2003}
\title{
  Small Exact D-, A-, G-, and IV-Optimal Designs
}
\description{
  Retrieve small exact D-, A-, G-, and IV-Optimal designs generated by a genetic algorithm. These designs are catalogued by Borkowski (2003).
}
\usage{
  Borkowski2003(criterion, k, N)
}
\arguments{
  \item{criterion}{
    Only optimal criteria "D", "A", "G", and "IV" are available now.
  }
  \item{k}{
    The number of factor, k = 2 and 3
    }
      \item{N}{
    The number of design points
    }
}
\value{
  \code{Borkowski2003} is called to generate the data frame of the design matrix of the optimal designs.
}
\examples{
  Borkowski2003("D", 2, 10)
  Borkowski2003("G", 3, 13)
  Borkowski2003("IV", 2, 7)
  Borkowski2003("A", 3, 15)
}
\references{
Borkowski, J. J. (2003). \emph{Using a Genetic Algorithm to Generate Small Exact Response Surface Designs}. Journal of Probability and Statistical Science, 1(1):65-88.
}
\keyword{Small Exact Optimal Designs}
