\name{VdgRsm}
\alias{VdgRsm}
\docType{package}
\title{
Variance Dispersion, Fraction of Design Space Plots, Designs, and Contour Plots
}
\description{
This package creates variance dispersion graphs (VDGs), fraction of design space (FDS) plots, and contour of SPVs for second-order model designs in spherical or cuboidal regions. The package also generates some response surface designs.
}
\details{\tabular{ll}{
Package: \tab VdgRsm\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2015-02-01\cr
License: \tab GPL-2
}}
\author{
Patchanok Srisuradetchai, John J Borkowski \cr
Maintainer: Patchanok Srisuradetchai <s.patchanok@gmail.com>
}
\references{
1. Borkowski J.J. (2003), "Using a Genetic Algorithm to Generate Small Exact Response Surface Designs", Journal of Probability and Statistical Science, 1(1):65-88.\cr
2. Borkowski J.J. (2005), Chapter 14: "Graphical Methods for Assessing the Prediction Capability of Response Surface Designs" In Khuri, A.I., "Response Surface Methodology and Related Topics", p.349-375, World Scientific Publishing.\cr
3. Doehlert, D. H (1970), "Uniform Shell Designs"", Journal of the Royal Statistical Society, 19(3):231-239. \cr
4. Giovannitti-Jensen, A. and Myers, R.H. (1989), "Graphical Assessment of the Prediction Capability of Response Surface Designs", Technometrics, 31, p.159-171.\cr
5. Khuri, A.I., Kim, H.J., and Um Y. (1996), "Quantile plots of the prediction variance for response surface designs", Computational Statistics and Data Analysis, 22, p.395-407. \cr
6. Nguyen, N.K. and Borkowski, J.J. (2008), "New 3-Level Response Surface Designs Constructed from Incomplete Block Designs", Journal of Statistical Planning and Inference, 138, p.294-305.\cr 
7. Rozum, M.A. and Myers, R.H. (1991), "Adaptation of Variance Dispersion Graphs to Cuboidal Regions of Interest", Presented at Joint Statistical Meetings, American Statistical Association, Atlanta, GA. \cr
8. SAS 9.1 ADX Interface for Design of Experiments. Cary, NC: SAS Institute Inc. \cr
9. Zahran, A., Anderson-Cook, C.M., and Myers, R.H. (2003), "Fraction of Design Space to Assess the Prediction Capability of Response Surface Designs", Journal of Quality Technology, 35, p.377-386.
}
\keyword{design}
\seealso{
The CRAN task view on Design of Experiments
}
